Update: Validation all Page Admin

+++ app/Http/Controllers/Admin/DivisionController.php
@@ -81,6 +81,11 @@ class DivisionController extends Controller
     public function destroy(string $id)
     {
         $division = Division::findOrFail($id);
+
+        if($division->users()->count() > 0){
+            Alert::error('Error', 'Data tidak bisa dihapus karena masih digunakan');
+            return redirect()->route('division.index');
+        }
         
         $division->delete();
